Update 18f-source-code-policy.md
Here's a bit of tinkering. Comments are in markdown. 

Following Eric's suggestion, I tried to make this a more human-oriented. I think this can serve as an internal "north star" as we think about how we expand this line of thinking to the rest of the agency. Welcome any and all suggestions for how to modify.