Fix link style
The old link style must have been for some kind of a renderer, like
Jekyll, that is no longer in use? It doesn't appear that this site is
published elsewhere as HTML. Linking just the number `4` doesn't make
any sense when reading through the rendered Markdown, so I propose
simply linking the text in question.